Backend/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
Deprecated==1.2.13
firebase==3.0.1
gcloud==0.18.3
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
jwcrypto==1.0
oauth2client==4.1.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
pyparsing==3.0.6
python-jwt==3.3.2
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sseclient==0.0.27
urllib3==1.26.8
wrapt==1.13.3
keyboard==0.13.5
opencv-python
keras
scikit-learn
pandas
pyrebase4
ar_corrector